IR control decoder task working.
The implementation of IR decoding is tested and working. Several things was needed for this to happens. There is an issue with configuring the GPIO interrupt for falling edge detection so the implementation currently uses rising edge detection and an inverter gate is in front of the GPIO to negate the sensor output before going into MCU DIO 4 pin. micros function is implemented using DualTimer module.